research/ReFound/requirements.txt
numpy
PIL
paddlepaddle-gpu==2.4.2
scikit-learn
scipy